+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/GameOpModes/HzAutonomousBasic.java
@@ -3,7 +3,6 @@ package org.firstinspires.ftc.teamcode.GameOpModes;
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.trajectory.Trajectory;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
@@ -22,19 +21,22 @@ import static com.qualcomm.robotcore.util.ElapsedTime.Resolution.MILLISECONDS;
 
 
 /**
- * This opmode demonstrates how one would implement field centric control using
- * `SampleMecanumDrive.java`. This file is essentially just `TeleOpDrive.java` with the addition of
- * field centric control. To achieve field centric control, the only modification one needs is to
- * rotate the input vector by the current heading before passing it into the inverse kinematics.
- * <p>
- * See lines 42-57.
+ * Ultimate Goal Autonomous mode <BR>
  *
- * setAutoMoveArmPickWobble()
- * setAutoMoveArmDropWobbleRing()
+ * This code describes how Autonomous mode is done by Hazmat Robot for Ultimate Goal.<BR>
+ * The following options are coded here, and selectable through Gamepad inputs to set up <BR>
+ *     <emsp>Playing Alliance : Red or Blue</emsp>
+ *     <emsp>Start Line : Inner or Outer</emsp>
+ *     <emsp>Game options :</emsp>
+ *     <emsp>     Launch rings to High Goal or Power Shot and park</emsp>
+ *     <emsp>     Launch rings to High Goal or Power Shot, drop Wobble Goal and park</emsp>
+ *     <emsp>     Launch rings to High Goal or Power Shot, drop Wobble Goal, Pick rings from target marker, launch and park</emsp>
+ *     <emsp>     Launch rings to High Goal or Power Shot, drop Wobble Goal, Pick rings from target marker, launch, move Wobble Goal2 and park</emsp>
  *
+ * The code for Red and Blue are written as reflection of each other.<BR>
+ * Camera on either side is used using Vuforia to determine target for Wobble Goal<BR>
  */
 @Autonomous(name = "Hazmat Autonomous", group = "00-Autonomous" , preselectTeleOp = "Hazmat TeleOp RR")
-@Disabled
 public class HzAutonomousBasic extends LinearOpMode {
 
     public boolean HzDEBUG_FLAG = true;
@@ -57,6 +59,8 @@ public class HzAutonomousBasic extends LinearOpMode {
     public HzGameField.TARGET_ZONE targetZone = HzGameField.TARGET_ZONE.A;
     public HzVuforia.ACTIVE_WEBCAM activeWebcam = HzVuforia.ACTIVE_WEBCAM.LEFT;
 
+    double af = HzGameField.ALLIANCE_FACTOR;
+
     double turnAnglePowershot12 = Math.toRadians(-5);
     double turnAnglePowershot23 = Math.toRadians(-5);
     Trajectory traj;
@@ -64,7 +68,7 @@ public class HzAutonomousBasic extends LinearOpMode {
 
     @Override
     public void runOpMode() throws InterruptedException {
-        // Initialize SampleMecanumDrive
+        // Initialize HW
         hzDrive = new HzDrive(hardwareMap);
         hzMagazine = new HzMagazine(hardwareMap);
         hzIntake = new HzIntake(hardwareMap);
@@ -75,13 +79,12 @@ public class HzAutonomousBasic extends LinearOpMode {
         hzGamepad = new HzGamepad(gamepad1,hzDrive,hzMagazine,hzIntake,hzLaunchController,hzLauncher,hzArm);
         hzAutoControl = new HzAutoControl(hzDrive,hzMagazine,hzIntake,hzLaunchController,hzLauncher,hzArm);
 
-        //initialConfiguration();
+        //Key Pay inputs to select Game Plan;
         selectGamePlan();
         hzVuforia = new HzVuforia(hardwareMap, activeWebcam);
+        af = HzGameField.ALLIANCE_FACTOR;
 
-        // Initiate Camera even before Start is pressed.
-        //waitForStart();
-
+        // Initiate Camera on Init.
         hzVuforia.activateVuforiaTensorFlow();
 
         hzDrive.getLocalizer().setPoseEstimate(startPose);
@@ -100,10 +103,13 @@ public class HzAutonomousBasic extends LinearOpMode {
         if (isStopRequested()) return;
 
         while (!isStopRequested()) {
-            //Init is pressed at this time, and start is not pressed yet
-
             //Run Vuforia Tensor Flow
             targetZone = hzVuforia.runVuforiaTensorFlow();
+
+            if (!parked){
+                hzAutoControl.setMagazineToLaunch();
+                hzAutoControl.runAutoControl();
+            }
 
             if (HzDEBUG_FLAG) {
                 printDebugMessages();
@@ -111,32 +117,24 @@ public class HzAutonomousBasic extends LinearOpMode {
             }
             hzLaunchController.launchMode = HzLaunchController.LAUNCH_MODE.MANUAL;
 
+            //Game Play is pressed
             while (opModeIsActive() && !isStopRequested() && !parked) {
 
                 hzVuforia.deactivateVuforiaTensorFlow();
 
-                if (HzGameField.playingAlliance == HzGameField.PLAYING_ALLIANCE.BLUE_ALLIANCE) {
-                    if (startPose == HzGameField.BLUE_INNER_START_LINE ) {
-                        autonomousStarted = true;
-                        runAutoBlueInner();
-                    } else if (startPose == HzGameField.BLUE_OUTER_START_LINE ) {
-                        autonomousStarted = true;
-                        runAutoBlueOuterHighGoal();
-                    }
-                } else { //HzGameField.playingAlliance == HzGameField.PLAYING_ALLIANCE.RED_ALLIANCE
-                    if (startPose == HzGameField.RED_INNER_START_LINE ) {
-                        autonomousStarted = true;
-                        //runAutoRedInnerHighGoal();
-                    } else if (startPose == HzGameField.RED_OUTER_START_LINE ) {
-                        autonomousStarted = true;
-                        //runAutoRedOuterHighGoal();
-                    }
+                if (HzGameField.startPosition == HzGameField.START_POSITION.INNER) {
+                    runAutoInner();
+                } else { //HzGameField.startPosition == HzGameField.START_POSITION.OUTER
+                    runAutoOuter();
                 }
 
                 hzIntake.setIntakeReleaseOpen();
+                hzAutoControl.setMagazineToCollect();
 
                 //Move to Launching Position
                 parked = true;
+
+                //Write last position to static class to be used as initial position in TeleOp
                 HzGameField.currentPose = hzDrive.getPoseEstimate();
                 HzGameField.poseSetInAutonomous = true;
 
@@ -148,194 +146,292 @@ public class HzAutonomousBasic extends LinearOpMode {
 
         }
 
-        // Transfer the current pose to PoseStorage so we can use it in TeleOp
+        //Write last position to static class to be used as initial position in TeleOp
         HzGameField.currentPose = hzDrive.getPoseEstimate();
         HzGameField.poseSetInAutonomous = true;
     }
 
-    public void runAutoBlueInner(){
+    /**
+     * Path and actions for autonomous mode starting from Inner start position
+     */
+    public void runAutoInner(){
 
         // Set magazine to Launch in case it slipped
         hzAutoControl.setMagazineToLaunch();
 
+        // For Launch and Park mode, wait time, so that alliance robot is not crossed
+        if (!hzAutoControl.dropFirstWobbleGoal){
+            hzWait(20000);
+        }
+
         // Move to launch position and launch rings to High Goal or Powershots
         if (hzAutoControl.autoLaunchAim == HzAutoControl.AutoLaunchAim.HIGHGOAL) {
             hzAutoControl.setLaunchTargetHighGoal();
-            traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                    .lineToLinearHeading(new Pose2d(-40,6,Math.toRadians(45)))
-                    .build();
-            hzDrive.followTrajectory(traj);
+
+            //Intermediary position to move away from alliance robot.
+            //For option of second wobble goal drop, this is avoided to save time
+            if (!hzAutoControl.pickAndDropSecondWobbleGoal) {
+                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                        .lineToLinearHeading(new Pose2d(-34, af * 14, Math.toRadians(af * 45))) //-40,6
+                        .build();
+                hzDrive.followTrajectory(traj);
+            }
             hzAutoControl.setMagazineToLaunch();
+            //Move to position to launch rings
             traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                    .lineToLinearHeading(new Pose2d(-10,14,Math.toRadians(18)))
+                    .lineToLinearHeading(new Pose2d(-10,af*14,Math.toRadians(af*17))) //18
                     .build();
             hzDrive.followTrajectory(traj);
 
             launch3RingsToHighGoal();
         } else {
             hzAutoControl.setLaunchTargetPowerShot1();
+
+            //Intermediary position to move away from alliance robot.
+            //For option of second wobble goal drop, this is avoided to save time
+            if (!hzAutoControl.pickAndDropSecondWobbleGoal) {
+                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                        .lineToLinearHeading(new Pose2d(-40, af * 6, Math.toRadians(af * 45)))
+                        .build();
+                hzDrive.followTrajectory(traj);
+            }
+            //Move to position to launch rings
             traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                    .lineToLinearHeading(new Pose2d(-40,6,Math.toRadians(45)))
-                    .build();
-            hzDrive.followTrajectory(traj);
-            traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                    .splineToLinearHeading(new Pose2d(-10,14,Math.toRadians(5)),Math.toRadians(0))
+                    .lineToLinearHeading(new Pose2d(-10,af*14,Math.toRadians(af*5)))
                     .build();
             hzDrive.followTrajectory(traj);
 
-            //Set turn angles prior to launching
-            turnAnglePowershot12 = Math.toRadians(-5);
-            turnAnglePowershot23 = Math.toRadians(-5);
+            //Set turn angles prior to launching for each of the power shorts
+            turnAnglePowershot12 = Math.toRadians(af*-5);
+            turnAnglePowershot23 = Math.toRadians(af*-5);
             launch3RingsToPowerShots();
         }
 
-        // Move to drop wobble goal on target
-        switch (targetZone){
-            case A:
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToSplineHeading(new Pose2d(46,22,Math.toRadians(-45)))
-                        .build();
-                hzDrive.followTrajectory(traj);
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToSplineHeading(new Pose2d(27,43,Math.toRadians(-45)))
-                        .build();
-                hzDrive.followTrajectory(traj);
-
-                break;
-
-            case B:
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToSplineHeading(new Pose2d(46,22,Math.toRadians(-45)))
-                        .build();
-                hzDrive.followTrajectory(traj);
-                break;
-            case C:
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToSplineHeading(new Pose2d(52,43,Math.toRadians(-90)))
-                        .build();
-                hzDrive.followTrajectory(traj);
-                break;
-
-        }
-
-        dropWobbleGoalInTarget();
-
-        if ((hzAutoControl.pickRingFromTargetMarker == false) || (targetZone == HzGameField.TARGET_ZONE.A)){ // Park
-            //Park
+        if (!hzAutoControl.dropFirstWobbleGoal){
+            //Park after launch rings
             traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                    .lineToLinearHeading(new Pose2d(50,16,Math.toRadians(-45)))
+                    .lineToSplineHeading(new Pose2d(13, af * 20, Math.toRadians(af * 0)))
                     .build();
             hzDrive.followTrajectory(traj);
+            return;
+        } else {
+            // Move to drop wobble goal on target
+            switch (targetZone) {
+                case A:
+                    //For option of second wobble goal drop, this intermediary position is avoided to save time
+                    /*if (!hzAutoControl.pickAndDropSecondWobbleGoal) {
+                        traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                                .lineToSplineHeading(new Pose2d(46, af * 22, Math.toRadians(af * -45)))
+                                .build();
+                        hzDrive.followTrajectory(traj);
+                    }*/
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToSplineHeading(new Pose2d(27, af * 48, Math.toRadians(af * -45)))//43
+                            .build();
+                    hzDrive.followTrajectory(traj);
 
-            traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                    .lineToSplineHeading(new Pose2d(13, 20, Math.toRadians(0)))
-                    .build();
-            hzDrive.followTrajectory(traj);
+                    break;
+
+                case B:
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToSplineHeading(new Pose2d(25, af* 22, Math.toRadians(af * -135)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
+                    break;
+                case C:
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToSplineHeading(new Pose2d(54, af * 43, Math.toRadians(af * -90))) //y:51
+                            .build();
+                    hzDrive.followTrajectory(traj);
+                    break;
+            }
+
+            dropWobbleGoalInTarget();
+
+            if ((hzAutoControl.pickRingFromTargetMarker == false) || (targetZone == HzGameField.TARGET_ZONE.A)) { // Park
+                //Park
+                if (hzAutoControl.pickAndDropSecondWobbleGoal) {
+                    runInnerPickAndDropSecondWobbleGoalAndPark();
+                } else {
+                    //Move towards base line away from other robot and Park
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToLinearHeading(new Pose2d(50, af * 16, Math.toRadians(af * -45)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToSplineHeading(new Pose2d(13, af * 20, Math.toRadians(af * 0)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
+                }
 
             /* Alternate
             traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                    .lineToSplineHeading(new Pose2d(10, 0, Math.toRadians(0)))
+                    .lineToSplineHeading(new Pose2d(10, af*0, Math.toRadians(0)))
                     .build();
             hzDrive.followTrajectory(traj);
             */
-            return;
-        } else { //Target Zone is B or C and pickRingFromTargetMarker == True
-            //Pick rings from Target Marker
-            hzIntake.setIntakeReleaseOpen();
-            hzAutoControl.setIntakeStart();
-            hzWait(500);
+                return;
+            } else { //Target Zone is B or C and pickRingFromTargetMarker == True
+                //Pick rings from Target Marker
+                hzIntake.setIntakeReleaseOpen();
+                hzAutoControl.setIntakeStart();
+                hzWait(500);
 
-            if (targetZone == HzGameField.TARGET_ZONE.B) {
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToLinearHeading(new Pose2d(-10,33,Math.toRadians(135)))
-                        .build();
-                hzDrive.followTrajectory(traj);
+                //Move to Position to pick rings
+                if (targetZone == HzGameField.TARGET_ZONE.B) {
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToLinearHeading(new Pose2d(-10, af * 36, Math.toRadians(af * 180)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
 
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToLinearHeading(new Pose2d(-22,33,Math.toRadians(135)))
-                        .build();
-                hzDrive.followTrajectory(traj);
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToLinearHeading(new Pose2d(-22, af * 36, Math.toRadians(af * 180)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
 
-                hzWait(300);
-            }
+                    hzWait(300);
+                }
 
-            // Spline to (24,24,0)
-            if (targetZone == HzGameField.TARGET_ZONE.C) {
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToLinearHeading(new Pose2d(-10,33,Math.toRadians(-180)))
-                        .build();
-                hzDrive.followTrajectory(traj);
+                // Spline to (24,24,0)
+                if (targetZone == HzGameField.TARGET_ZONE.C) {
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToLinearHeading(new Pose2d(-10, af * 36, Math.toRadians(af * -180)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
 
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToLinearHeading(new Pose2d(-22,33,Math.toRadians(-180)))
-                        .build();
-                hzDrive.followTrajectory(traj);
+                    /*if (!hzAutoControl.pickAndDropSecondWobbleGoal) {
+                        traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                                .lineToLinearHeading(new Pose2d(-22, af * 37, Math.toRadians(af * -180)))
+                                .build();
+                        hzDrive.followTrajectory(traj);
+                    }*/
 
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToLinearHeading(new Pose2d(-33, 33, Math.toRadians(-180)))
-                        .build();
-                hzDrive.followTrajectory(traj);
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToLinearHeading(new Pose2d(-27, af * 36, Math.toRadians(af * -180))) //x33
+                            .build();
+                    hzDrive.followTrajectory(traj);
 
-                hzWait(300);
-            }
+                    hzWait(300);
+                }
 
-            if (hzAutoControl.launchRingsPickedFromTargetMarkerToHighGoal) {
-                hzAutoControl.setIntakeStop();
-                hzAutoControl.setMagazineToLaunch();
-                hzAutoControl.setLaunchTargetHighGoal();
+                if (hzAutoControl.launchRingsPickedFromTargetMarkerToHighGoal) {
+                    hzAutoControl.setIntakeStop();
+                    hzAutoControl.setMagazineToLaunch();
+                    hzAutoControl.setLaunchTargetHighGoal();
 
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToLinearHeading(new Pose2d(-13, 33, Math.toRadians(0)))
-                        .build();
-                hzDrive.followTrajectory(traj);
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToLinearHeading(new Pose2d(-13, af * 40, Math.toRadians(af * 3)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
 
-                launch3RingsToHighGoal();
+                    launch3RingsToHighGoal();
 
-                //Park
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToSplineHeading(new Pose2d(13, 33, Math.toRadians(0)))
-                        .build();
-                hzDrive.followTrajectory(traj);
-            } else { //Move to baseline and park in safe zone
-                traj =  hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToLinearHeading(new Pose2d(-50,16,Math.toRadians(-90)))
-                        .build();
-                hzDrive.followTrajectory(traj);
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToSplineHeading(new Pose2d(10, 0, Math.toRadians(0)))
-                        .build();
-                hzDrive.followTrajectory(traj);
-                hzAutoControl.setIntakeStop();
+                    if (hzAutoControl.pickAndDropSecondWobbleGoal == true) {
+                        runInnerPickAndDropSecondWobbleGoalAndPark();
+                    } else {
+                        //Park
+                        traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                                .lineToSplineHeading(new Pose2d(13, af * 33, Math.toRadians(af * 0)))
+                                .build();
+                        hzDrive.followTrajectory(traj);
+                    }
+                } else { //Move to baseline and park in safe zone
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToLinearHeading(new Pose2d(-50, af * 16, Math.toRadians(af * -90)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToSplineHeading(new Pose2d(10, af * 0, Math.toRadians(af * 0)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
+                    hzAutoControl.setIntakeStop();
+                }
             }
         }
     }
 
-    public void runAutoBlueOuterHighGoal(){
+    /**
+     * Path for picking and dropping second wobble goal for Inner Start position
+     */
+    public void runInnerPickAndDropSecondWobbleGoalAndPark() {
+        hzAutoControl.setMoveArmPickWobble();
+        hzAutoControl.runOpenGrip();
+        traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                .lineToLinearHeading(new Pose2d(-33, af*50, Math.toRadians(af*0)))
+                .build();
+        hzDrive.followTrajectory(traj);
+        hzWait(200);
+        hzAutoControl.runCloseGrip();
+        hzWait(400);
+        hzAutoControl.setMoveArmHoldUpWobbleRing();
+        hzWait(500);
+        switch (targetZone) {
+            case A:
+                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                        .lineToSplineHeading(new Pose2d(13, af*20, Math.toRadians(af*-60)))
+                        .build();
+                hzDrive.followTrajectory(traj);
+                break;
+            case B:
+                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                        .lineToSplineHeading(new Pose2d(13, af*33, Math.toRadians(af*180)))
+                        .build();
+                hzDrive.followTrajectory(traj);
+                break;
+            case C:
+                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                        .lineToSplineHeading(new Pose2d(42, af*50, Math.toRadians(af*-135)))
+                        .build();
+                hzDrive.followTrajectory(traj);
+                break;
+        }
+        dropWobbleGoalInTarget();
+        //Park
+        if (targetZone == HzGameField.TARGET_ZONE.C) {
+            traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                    .lineToSplineHeading(new Pose2d(13, af * 33, Math.toRadians(af * 0)))
+                    .build();
+            hzDrive.followTrajectory(traj);
+        }
+    }
+
+    /**
+     * Path and actions for autonomous mode starting from Outer start position
+     */
+    public void runAutoOuter(){
         // Set magazine to Launch in case it slipped
         hzAutoControl.setMagazineToLaunch();
+
+        if (!hzAutoControl.dropFirstWobbleGoal){
+            hzWait(20000);
+        }
 
         // Move to launch position and launch rings to High Goal or Powershots
         if (hzAutoControl.autoLaunchAim == HzAutoControl.AutoLaunchAim.HIGHGOAL) {
             hzAutoControl.setLaunchTargetHighGoal();
-            traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                    .lineToLinearHeading(new Pose2d(-40,50,Math.toRadians(-45)))
-                    .build();
-            hzDrive.followTrajectory(traj);
+            if (!hzAutoControl.pickAndDropSecondWobbleGoal) {
+                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                        .lineToLinearHeading(new Pose2d(-40, af * 50, Math.toRadians(af * -45)))
+                        .build();
+                hzDrive.followTrajectory(traj);
+            }
             hzAutoControl.setMagazineToLaunch();
             traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                    .lineToLinearHeading(new Pose2d(-10,50,Math.toRadians(-10)))
+                    .lineToLinearHeading(new Pose2d(-10,af*50,Math.toRadians(af*-10)))
                     .build();
             hzDrive.followTrajectory(traj);
 
             launch3RingsToHighGoal();
         } else {
             hzAutoControl.setLaunchTargetPowerShot1();
+            if (!hzAutoControl.pickAndDropSecondWobbleGoal) {
+                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                        .lineToLinearHeading(new Pose2d(-40, af * 50, Math.toRadians(af * 45)))
+                        .build();
+                hzDrive.followTrajectory(traj);
+            }
             traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                    .lineToLinearHeading(new Pose2d(-40,50,Math.toRadians(45)))
-                    .build();
-            hzDrive.followTrajectory(traj);
-            traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                    .splineToLinearHeading(new Pose2d(-10,50,Math.toRadians(-15)),Math.toRadians(0))
+                    .splineToLinearHeading(new Pose2d(-10,af*50,Math.toRadians(af*-15)),Math.toRadians(0))
                     .build();
             hzDrive.followTrajectory(traj);
 
@@ -345,147 +441,223 @@ public class HzAutonomousBasic extends LinearOpMode {
             launch3RingsToPowerShots();
         }
 
-
-        switch (targetZone){
-            case A:
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToLinearHeading(new Pose2d(-10,48,Math.toRadians(-135)))
-                        .build();
-                hzDrive.followTrajectory(traj);
-                //Consider just turn(Math.toRadians(-120);
-                break;
-            case B:
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToSplineHeading(new Pose2d(15,48,Math.toRadians(135)))
-                        .build();
-                hzDrive.followTrajectory(traj);
-                break;
-            case C:
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToSplineHeading(new Pose2d(35,50,Math.toRadians(-135)))
-                        .build();
-                hzDrive.followTrajectory(traj);
-                break;
-
-        }
-
-        dropWobbleGoalInTarget();
-
-        if ((hzAutoControl.pickRingFromTargetMarker == false) || (targetZone == HzGameField.TARGET_ZONE.A)){ // Park
+        if (!hzAutoControl.dropFirstWobbleGoal) {
             //Park
             traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                    .lineToSplineHeading(new Pose2d(13, 20, Math.toRadians(0)))
+                    .lineToSplineHeading(new Pose2d(13, af * 40, Math.toRadians(af * 0)))
                     .build();
             hzDrive.followTrajectory(traj);
+            return;
+        } else {
+
+            switch (targetZone) {
+                case A:
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToLinearHeading(new Pose2d(-10, af * 48, Math.toRadians(af * -135)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
+                    break;
+                case B:
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToSplineHeading(new Pose2d(15, af * 48, Math.toRadians(af * 135)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
+                    break;
+                case C:
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToSplineHeading(new Pose2d(35, af * 50, Math.toRadians(af * -150)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
+                    break;
+
+            }
+
+            dropWobbleGoalInTarget();
+
+            if ((hzAutoControl.pickRingFromTargetMarker == false) || (targetZone == HzGameField.TARGET_ZONE.A)) { // Park
+                if (hzAutoControl.pickAndDropSecondWobbleGoal) {
+                    runOuterPickAndDropSecondWobbleGoalAndPark();
+                } else {
+                    //Park
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToSplineHeading(new Pose2d(13, af * 20, Math.toRadians(af * 0)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
 
             /* Alternate
             traj =  hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                    .lineToLinearHeading(new Pose2d(-50,16,Math.toRadians(-90)))
+                    .lineToLinearHeading(new Pose2d(-50,af*16,Math.toRadians(af*-90)))
                     .build();
             hzDrive.followTrajectory(traj);
             traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                    .lineToSplineHeading(new Pose2d(10, 0, Math.toRadians(0)))
+                    .lineToSplineHeading(new Pose2d(10, af*0, Math.toRadians(af*0)))
                     .build();
             hzDrive.followTrajectory(traj);
             */
-            return;
+                }
+                return;
 
-        } else { //Target Zone is B or C and pickRingFromTargetMarker == True
-            //Pick rings from Target Marker
-            hzIntake.setIntakeReleaseOpen();
-            hzAutoControl.setIntakeStart();
-            hzWait(500);
+            } else { //Target Zone is B or C and pickRingFromTargetMarker == True
+                //Pick rings from Target Marker
+                hzIntake.setIntakeReleaseOpen();
+                hzAutoControl.setIntakeStart();
+                hzWait(500);
 
-            if (targetZone == HzGameField.TARGET_ZONE.B) {
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToLinearHeading(new Pose2d(-10,33,Math.toRadians(135)))
-                        .build();
-                hzDrive.followTrajectory(traj);
+                if (targetZone == HzGameField.TARGET_ZONE.B) {
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToLinearHeading(new Pose2d(-10, af * 33, Math.toRadians(af * 135)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
 
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToLinearHeading(new Pose2d(-22,33,Math.toRadians(135)))
-                        .build();
-                hzDrive.followTrajectory(traj);
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToLinearHeading(new Pose2d(-22, af * 33, Math.toRadians(af * 135)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
 
-                hzWait(300);
-            }
+                    hzWait(300);
+                }
 
-            // Spline to (24,24,0)
-            if (targetZone == HzGameField.TARGET_ZONE.C) {
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToLinearHeading(new Pose2d(-10,33,Math.toRadians(-180)))
-                        .build();
-                hzDrive.followTrajectory(traj);
+                // Spline to (24,24,0)
+                if (targetZone == HzGameField.TARGET_ZONE.C) {
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToLinearHeading(new Pose2d(-10, af * 36, Math.toRadians(af * -180)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
 
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToLinearHeading(new Pose2d(-22,33,Math.toRadians(-180)))
-                        .build();
-                hzDrive.followTrajectory(traj);
+                    /*if (!hzAutoControl.pickAndDropSecondWobbleGoal) {
+                        traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                                .lineToLinearHeading(new Pose2d(-22, af * 33, Math.toRadians(af * -180)))
+                                .build();
+                        hzDrive.followTrajectory(traj);
+                    }*/
 
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToLinearHeading(new Pose2d(-33, 33, Math.toRadians(-180)))
-                        .build();
-                hzDrive.followTrajectory(traj);
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToLinearHeading(new Pose2d(-33, af * 36, Math.toRadians(af * -180)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
 
-                hzWait(300);
-            }
+                    hzWait(300);
+                }
 
-            if (hzAutoControl.launchRingsPickedFromTargetMarkerToHighGoal) {
-                hzAutoControl.setIntakeStop();
-                hzAutoControl.setMagazineToLaunch();
-                hzAutoControl.setLaunchTargetHighGoal();
+                if (hzAutoControl.launchRingsPickedFromTargetMarkerToHighGoal) {
+                    hzAutoControl.setIntakeStop();
+                    hzAutoControl.setLaunchTargetHighGoal();
+                    hzAutoControl.setMagazineToLaunch();
 
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToLinearHeading(new Pose2d(-13, 33, Math.toRadians(0)))
-                        .build();
-                hzDrive.followTrajectory(traj);
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToLinearHeading(new Pose2d(-13, af * 36, Math.toRadians(af * 0)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
 
-                launch3RingsToHighGoal();
+                    launch3RingsToHighGoal();
 
-                //Park
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToSplineHeading(new Pose2d(13, 33, Math.toRadians(0)))
-                        .build();
-                hzDrive.followTrajectory(traj);
-            } else { //Move to baseline and park in safe zone
-                traj =  hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToLinearHeading(new Pose2d(-50,16,Math.toRadians(-90)))
-                        .build();
-                hzDrive.followTrajectory(traj);
-                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
-                        .lineToSplineHeading(new Pose2d(10, 0, Math.toRadians(0)))
-                        .build();
-                hzDrive.followTrajectory(traj);
-                hzAutoControl.setIntakeStop();
+                    if (hzAutoControl.pickAndDropSecondWobbleGoal == true) {
+                        runOuterPickAndDropSecondWobbleGoalAndPark();
+                    } else {
+
+                        //Park
+                        traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                                .lineToSplineHeading(new Pose2d(13, af * 33, Math.toRadians(af * 0)))
+                                .build();
+                        hzDrive.followTrajectory(traj);
+                    }
+                } else { //Move to baseline and park in safe zone
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToLinearHeading(new Pose2d(-50, af * 16, Math.toRadians(af * -90)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
+                    traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                            .lineToSplineHeading(new Pose2d(10, af * 0, Math.toRadians(af * 0)))
+                            .build();
+                    hzDrive.followTrajectory(traj);
+                    hzAutoControl.setIntakeStop();
+                }
             }
         }
     }
 
-    public void launch3RingsToHighGoal(){
-        hzAutoControl.setMagazineToLaunch();
+    /**
+     * Path for picking and dropping second wobble goal for Outer Start position
+     */
+    public void runOuterPickAndDropSecondWobbleGoalAndPark() {
+        hzAutoControl.setMoveArmPickWobble();
+        hzAutoControl.runOpenGrip();
+        traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                .lineToLinearHeading(new Pose2d(-31, af*23, Math.toRadians(af*0)))
+                .build();
+        hzDrive.followTrajectory(traj);
+        hzWait(200);
+        hzAutoControl.runCloseGrip();
+        hzWait(400);
+        hzAutoControl.setMoveArmHoldUpWobbleRing();
         hzWait(500);
+        switch (targetZone) {
+            case A:
+                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                        .lineToSplineHeading(new Pose2d(13, af*20, Math.toRadians(af*-60)))
+                        .build();
+                hzDrive.followTrajectory(traj);
+                break;
+            case B:
+                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                        .lineToSplineHeading(new Pose2d(13, af*33, Math.toRadians(af*180)))
+                        .build();
+                hzDrive.followTrajectory(traj);
+                break;
+            case C:
+                traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                        .lineToSplineHeading(new Pose2d(37, af*50, Math.toRadians(af*-135)))
+                        .build();
+                hzDrive.followTrajectory(traj);
+                break;
+        }
+        dropWobbleGoalInTarget();
+        //Park
+        if (targetZone == HzGameField.TARGET_ZONE.C) {
+            traj = hzDrive.trajectoryBuilder(hzDrive.getPoseEstimate())
+                    .lineToSplineHeading(new Pose2d(13, af * 33, Math.toRadians(af * 0)))
+                    .build();
+            hzDrive.followTrajectory(traj);
+        }
+    }
+
+    /**
+     * Sequence of launching 3 rings to High Goal
+     */
+    public void launch3RingsToHighGoal(){
         hzAutoControl.setLaunchTargetHighGoal();
         hzAutoControl.setMagazineToLaunch();
-        hzWait(400);
+        hzWait(200);
+        hzAutoControl.setLaunchTargetHighGoal();
+        hzAutoControl.setMagazineToLaunch();
+        hzWait(450); //400
         hzAutoControl.setRunLauncherTrue();
         hzAutoControl.setMagazineToLaunch();
-        hzWait(400);
+        hzWait(450);
         hzAutoControl.setRunLauncherTrue();
         hzAutoControl.setMagazineToLaunch();
-        hzWait(400);
+        hzWait(450);
         hzAutoControl.setRunLauncherTrue();
-        hzWait(400);
+
+        // Run 4th just in case
+        if (!hzAutoControl.pickAndDropSecondWobbleGoal) {
+            hzAutoControl.setMagazineToLaunch();
+            hzWait(450);
+            hzAutoControl.setRunLauncherTrue();
+        }
+        hzWait(600);
 
         hzAutoControl.setLaunchTargetOff();
         hzAutoControl.setMagazineToCollect();
     }
 
+    /**
+     * Sequence of launching 3 rings to PowerShots
+     */
     public void launch3RingsToPowerShots(){
-        hzAutoControl.setMagazineToLaunch();
-        hzWait(500);
         hzAutoControl.setLaunchTargetPowerShot1();
         hzAutoControl.setMagazineToLaunch();
-        hzWait(400);
+        hzWait(600);
         hzAutoControl.setRunLauncherTrue();
 
         hzDrive.turn(turnAnglePowershot12);
@@ -499,15 +671,20 @@ public class HzAutonomousBasic extends LinearOpMode {
         hzAutoControl.setLaunchTargetPowerShot1();
         hzWait(400);
         hzAutoControl.setRunLauncherTrue();
+        hzWait(400);
 
         hzAutoControl.setLaunchTargetOff();
         hzAutoControl.setMagazineToCollect();
 
     }
 
+    /**
+     * Sequence of dropping Wobble Goal once it reaches target position
+     */
     public void dropWobbleGoalInTarget(){
         hzAutoControl.setMoveArmDropWobbleAutonoumous();
-        hzWait(1000);
+        //hzWait(1000);
+        hzWait(500);
         hzAutoControl.runOpenGrip();
         hzWait(300);
         hzAutoControl.setMoveArmHoldUpWobbleRing();
@@ -515,6 +692,10 @@ public class HzAutonomousBasic extends LinearOpMode {
     }
 
 
+    /**
+     * Safe method to wait so that stop button is also not missed
+     * @param time time in ms to wait
+     */
     public void hzWait(double time){
         ElapsedTime timer = new ElapsedTime(MILLISECONDS);
         timer.reset();
@@ -523,92 +704,25 @@ public class HzAutonomousBasic extends LinearOpMode {
         }
     }
 
-
-    public void initialConfiguration(){
-        telemetry.setAutoClear(true);
-        telemetry.addData("Compile time : ", "6:50 : 12/05");
-
-        //***** Select Alliance ******
-        telemetry.addData("Enter PLaying Alliance :", "(Red:B, Blue:X)");
-        telemetry.update();
-
-        while (!isStopRequested()) {
-            if (hzGamepad.getButtonBPress()) {
-                HzGameField.playingAlliance = HzGameField.PLAYING_ALLIANCE.RED_ALLIANCE;
-                telemetry.addData("Playing Alliance Selected : ", "RED_ALLIANCE");
-                break;
-            }
-            if (hzGamepad.getButtonXPress()) {
-                HzGameField.playingAlliance = HzGameField.PLAYING_ALLIANCE.BLUE_ALLIANCE;
-                telemetry.addData("Playing Alliance Selected : ", "BLUE_ALLIANCE");
-                break;
-            }
-            //telemetry.addData("10s time : Default Alliance A :",);
-            telemetry.update();
-        }
-
-        telemetry.update();
-        sleep(500);
-
-        //***** Select Start Pose ******
-        //timer.reset();
-        telemetry.addData("Enter Start Pose :", "(Inner:A, Outer:Y)");
-        while (!isStopRequested()) {
-            if (HzGameField.playingAlliance == HzGameField.PLAYING_ALLIANCE.RED_ALLIANCE) {
-                if (hzGamepad.getButtonAPress()) {
-                    startPose = HzGameField.RED_INNER_START_LINE;
-                    activeWebcam = HzVuforia.ACTIVE_WEBCAM.LEFT;
-                    telemetry.addData("Start Pose : ", "RED_INNER_START_LINE");
-                    break;
-                }
-                if (hzGamepad.getButtonYPress()) {
-                    startPose = HzGameField.RED_OUTER_START_LINE;
-                    activeWebcam = HzVuforia.ACTIVE_WEBCAM.RIGHT;
-                    telemetry.addData("Start Pose : ", "RED_OUTER_START_LINE");
-                    break;
-                }
-            }
-            if (HzGameField.playingAlliance == HzGameField.PLAYING_ALLIANCE.BLUE_ALLIANCE) {
-                if (hzGamepad.getButtonAPress()) {
-                    startPose = HzGameField.BLUE_INNER_START_LINE;
-                    activeWebcam = HzVuforia.ACTIVE_WEBCAM.RIGHT;
-                    telemetry.addData("Start Pose : ", "BLUE_INNER_START_LINE");
-                    break;
-                }
-                if (hzGamepad.getButtonYPress()) {
-                    startPose = HzGameField.BLUE_OUTER_START_LINE;
-                    activeWebcam = HzVuforia.ACTIVE_WEBCAM.LEFT;
-                    telemetry.addData("Start Pose : ", "BLUE_OUTER_START_LINE");
-                    break;
-                }
-            }
-            //telemetry.addData("10s Timer : Default Pose : BLUE_INNER_START_LINE : %.3f", timer.time());
-            telemetry.update();
-        }
-        telemetry.update();
-        //sleep(500);
-    }
-
     public void selectGamePlan(){
         telemetry.setAutoClear(true);
-        telemetry.addData("Compile time : ", "4:11 :: 1/14");
+        telemetry.addData("Compile time : ", "4:47 :: 2/13");
 
         //***** Select Alliance ******
         telemetry.addData("Enter PLaying Alliance :", "(Blue: (X),    Red: (B))");
         telemetry.update();
 
-        /*HzGameField.playingAlliance = HzGameField.PLAYING_ALLIANCE.BLUE_ALLIANCE;
-        startPose = HzGameField.BLUE_INNER_START_LINE;
-        activeWebcam = HzVuforia.ACTIVE_WEBCAM.RIGHT;*/
 
         while (!isStopRequested()) {
             if (hzGamepad.getButtonBPress()) {
                 HzGameField.playingAlliance = HzGameField.PLAYING_ALLIANCE.RED_ALLIANCE;
+                HzGameField.ALLIANCE_FACTOR = -1;
                 telemetry.addData("Playing Alliance Selected : ", "RED_ALLIANCE");
                 break;
             }
             if (hzGamepad.getButtonXPress()) {
                 HzGameField.playingAlliance = HzGameField.PLAYING_ALLIANCE.BLUE_ALLIANCE;
+                HzGameField.ALLIANCE_FACTOR = 1;
                 telemetry.addData("Playing Alliance Selected : ", "BLUE_ALLIANCE");
                 break;
             }
@@ -619,17 +733,18 @@ public class HzAutonomousBasic extends LinearOpMode {
         hzWait(200);
 
         //***** Select Start Pose ******
-        //timer.reset();
         telemetry.addData("Enter Start Pose :", "(Inner: (A) ,    Outer: (Y))");
         while (!isStopRequested()) {
             if (HzGameField.playingAlliance == HzGameField.PLAYING_ALLIANCE.RED_ALLIANCE) {
                 if (hzGamepad.getButtonAPress()) {
+                    HzGameField.startPosition = HzGameField.START_POSITION.INNER;
                     startPose = HzGameField.RED_INNER_START_LINE;
                     activeWebcam = HzVuforia.ACTIVE_WEBCAM.LEFT;
                     telemetry.addData("Start Pose : ", "RED_INNER_START_LINE");
                     break;
                 }
                 if (hzGamepad.getButtonYPress()) {
+                    HzGameField.startPosition = HzGameField.START_POSITION.OUTER;
                     startPose = HzGameField.RED_OUTER_START_LINE;
                     activeWebcam = HzVuforia.ACTIVE_WEBCAM.RIGHT;
                     telemetry.addData("Start Pose : ", "RED_OUTER_START_LINE");
@@ -638,19 +753,20 @@ public class HzAutonomousBasic extends LinearOpMode {
             }
             if (HzGameField.playingAlliance == HzGameField.PLAYING_ALLIANCE.BLUE_ALLIANCE) {
                 if (hzGamepad.getButtonAPress()) {
+                    HzGameField.startPosition = HzGameField.START_POSITION.INNER;
                     startPose = HzGameField.BLUE_INNER_START_LINE;
                     activeWebcam = HzVuforia.ACTIVE_WEBCAM.RIGHT;
                     telemetry.addData("Start Pose : ", "BLUE_INNER_START_LINE");
                     break;
                 }
                 if (hzGamepad.getButtonYPress()) {
+                    HzGameField.startPosition = HzGameField.START_POSITION.OUTER;
                     startPose = HzGameField.BLUE_OUTER_START_LINE;
                     activeWebcam = HzVuforia.ACTIVE_WEBCAM.LEFT;
                     telemetry.addData("Start Pose : ", "BLUE_OUTER_START_LINE");
                     break;
                 }
             }
-            //telemetry.addData("10s Timer : Default Pose : BLUE_INNER_START_LINE : %.3f", timer.time());
             telemetry.update();
         }
         telemetry.update();
@@ -683,29 +799,53 @@ public class HzAutonomousBasic extends LinearOpMode {
             telemetry.addData("hzAutoControl.autoLaunchAim : ", hzAutoControl.autoLaunchAim);
             telemetry.addLine();
             telemetry.addData("Please select option for Picking rings from target markers and launching : ", "");
-            telemetry.addData("    (X):","Pick rings and launch after Wobble Goal, and park");
-            telemetry.addData("    (Y):","Pick rings after Wobble Goal and park (No launching)");
-            telemetry.addData("    (A):","Park after Wobble Goal (No ring Pick)");
+            telemetry.addData("    (Y):","Launch and park");
+            telemetry.addData("    (A):","Launch, drop WG and park");
+            telemetry.addData("    (X):","Launch, drop WG, Pick rings, launch and park");
+            telemetry.addData("    (B):","Launch, drop WG, Pick rings, launch, move WG2 and park");
 
-            if (hzGamepad.getButtonAPress()) {
+            if (hzGamepad.getButtonYPress()) {
+                hzAutoControl.dropFirstWobbleGoal = false;
                 hzAutoControl.pickRingFromTargetMarker = false;
                 hzAutoControl.launchRingsPickedFromTargetMarkerToHighGoal = false;
+                hzAutoControl.pickAndDropSecondWobbleGoal = false;
                 telemetry.addData("hzAutoControl.pickRingFromTargetMarker : ", hzAutoControl.pickRingFromTargetMarker);
                 telemetry.addData("hzAutoControl.launchRingsPickedFromTargetMarkerToHighGoal  : ", hzAutoControl.launchRingsPickedFromTargetMarkerToHighGoal);
+                telemetry.addData("hzAutoControl.pickAndDropSecondWobbleGoal: ", hzAutoControl.pickAndDropSecondWobbleGoal);
                 break;
             }
-            if (hzGamepad.getButtonYPress()) {
-                hzAutoControl.pickRingFromTargetMarker = true;
+
+            if (hzGamepad.getButtonAPress()) {
+                hzAutoControl.dropFirstWobbleGoal = true;
+                hzAutoControl.pickRingFromTargetMarker = false;
                 hzAutoControl.launchRingsPickedFromTargetMarkerToHighGoal = false;
+                hzAutoControl.pickAndDropSecondWobbleGoal = false;
+                telemetry.addData("hzAutoControl.dropFirstWobbleGoal : ", hzAutoControl.dropFirstWobbleGoal);
                 telemetry.addData("hzAutoControl.pickRingFromTargetMarker : ", hzAutoControl.pickRingFromTargetMarker);
                 telemetry.addData("hzAutoControl.launchRingsPickedFromTargetMarkerToHighGoal  : ", hzAutoControl.launchRingsPickedFromTargetMarkerToHighGoal);
+                telemetry.addData("hzAutoControl.pickAndDropSecondWobbleGoal: ", hzAutoControl.pickAndDropSecondWobbleGoal);
                 break;
             }
             if (hzGamepad.getButtonXPress()) {
+                hzAutoControl.dropFirstWobbleGoal = true;
                 hzAutoControl.pickRingFromTargetMarker = true;
                 hzAutoControl.launchRingsPickedFromTargetMarkerToHighGoal = true;
+                hzAutoControl.pickAndDropSecondWobbleGoal = false;
+                telemetry.addData("hzAutoControl.dropFirstWobbleGoal : ", hzAutoControl.dropFirstWobbleGoal);
                 telemetry.addData("hzAutoControl.pickRingFromTargetMarker : ", hzAutoControl.pickRingFromTargetMarker);
                 telemetry.addData("hzAutoControl.launchRingsPickedFromTargetMarkerToHighGoal  : ", hzAutoControl.launchRingsPickedFromTargetMarkerToHighGoal);
+                telemetry.addData("hzAutoControl.pickAndDropSecondWobbleGoal: ", hzAutoControl.pickAndDropSecondWobbleGoal);
+                break;
+            }
+            if (hzGamepad.getButtonBPress()) {
+                hzAutoControl.dropFirstWobbleGoal = true;
+                hzAutoControl.pickRingFromTargetMarker = true;
+                hzAutoControl.launchRingsPickedFromTargetMarkerToHighGoal = true;
+                hzAutoControl.pickAndDropSecondWobbleGoal = true;
+                telemetry.addData("hzAutoControl.dropFirstWobbleGoal : ", hzAutoControl.dropFirstWobbleGoal);
+                telemetry.addData("hzAutoControl.pickRingFromTargetMarker : ", hzAutoControl.pickRingFromTargetMarker);
+                telemetry.addData("hzAutoControl.launchRingsPickedFromTargetMarkerToHighGoal  : ", hzAutoControl.launchRingsPickedFromTargetMarkerToHighGoal);
+                telemetry.addData("hzAutoControl.pickAndDropSecondWobbleGoal: ", hzAutoControl.pickAndDropSecondWobbleGoal);
                 break;
             }
             telemetry.update();
@@ -726,6 +866,7 @@ public class HzAutonomousBasic extends LinearOpMode {
 
         telemetry.addData("GameField.playingAlliance : ", HzGameField.playingAlliance);
         telemetry.addData("startPose : ", startPose);
+        telemetry.addData("HzGameField.ALLIANCE_FACTOR",HzGameField.ALLIANCE_FACTOR);
 
         //****** Drive debug ******
         telemetry.addData("Drive Mode : ", hzDrive.driveMode);
@@ -804,13 +945,6 @@ public class HzAutonomousBasic extends LinearOpMode {
             }
         }
 
-
-
-
-        //***** Arm Debug ****
-        //telemetry.addData("armMotor.getTargetPosition()", hzArm.armMotor.getTargetPosition());
-        //telemetry.addData("armMotor.getCurrentPosition()", hzArm.armMotor.getCurrentPosition());
-
         switch (hzArm.getGripServoState()){
             case OPENED  : {
                 telemetry.addData("hzArm.getGripServoState()", "OPENED");
@@ -821,12 +955,6 @@ public class HzAutonomousBasic extends LinearOpMode {
                 break;
             }
         }
-
-        //telemetry.addData("armMotor.getCurrentPosition()", hzArm.armMotor.getCurrentPosition());
-        //telemetry.addData("armMotor.getTargetPosition()", hzArm.armMotor.getTargetPosition());
-
-        //telemetry.addData("armGripServo.getCurrentPosition()", hzArm.armGripServo.getPosition());
-        //telemetry.addData("hzGamepad.getLeftTriggerPress()", hzGamepad.getLeftTriggerPress());
 
         switch (hzArm.getCurrentArmPosition()){
             case PARKED: {
@@ -856,50 +984,3 @@ public class HzAutonomousBasic extends LinearOpMode {
     }
 }
 
-/*
-Turn command : drive.turn(Math.toRadians(ANGLE));
-
-Blue Alliance - Inner position-Target A
-    Start Pose : (TBD, 25, ~-70deg)
-    Spline to (0,12,0)
-    Launch ring to power shot 1,2,3, (automatic alignment)
-    Spline to (12,36,-90)
-    Drop wobble goal
-    Spline to (24,24,0)
-
-Blue Alliance - Inner position-Target B
-    Start Pose : (TBD, 25, ~-70deg)
-    Launch ring to power shot 1,2,3, (automatic alignment)
-    spline to (12,36,180)
-    Drop wobble goal
-    Turn intake on
-    Straight to (-24,36,180)
-    Turn 180 deg
-    Launch to high goal
-    Move to (12,36,0)
-
-Blue Alliance - Inner position-Target C - Option 1
-    Start Pose : (TBD, 25, ~-70deg)
-    Launch ring to power shot 1,2,3, (automatic alignment)
-    Turn  to 145 deg
-    Turn Intake on
-    Straight to (-24,36,145)
-    Turn to 0
-    Launch to High goal x 3
-    Spline to (40,40,-145)
-    Drop Wobble goal
-    Spline to (24,24,0)
-
-Blue Alliance - Inner position-Target C - Option 2
-    Start Pose : (TBD, 25, ~-70deg)
-    Launch ring to power shot 1,2,3, (automatic alignment)
-    Spline to (40,40,-145)
-    Drop Wobble goal
-    Turn Intake on
-    Spline to (-24,36,180)
-    Turn to 0
-    Launch to High goal x 3
-    Spline to (24,24,0)
-
-
- */
