+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/SubSystems/HzIntake.java
@@ -4,6 +4,22 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.hardware.Servo;
 
+/**
+ * Definition of HzIntake Class <BR>
+ *
+ * HzIntake consists of system provided intake controls and adds functionality to the selection made on intake. <BR>
+ *
+ * The states are as followed: <BR>
+ *     <emsp>INTAKE_MOTOR_STATE for telling if intake is running, stopped, or reversing </emsp> <BR>
+ *     <emsp>INTAKE_REVERSE_BUTTON_STATE sees if the intake is on or off </emsp> <BR>
+ *
+ * The functions are as followed: <BR>
+ *     <emsp>runIntakeMotor checks if the intake is not running and runs the intake </emsp> <BR>
+ *     <emsp>stopIntakeMotor checks if the intake has stopped and if its not, it sets the intake power to 0
+ *     and sets intakeMotorState to INTAKE_MOTOR_STATE.STOPPED </emsp> <BR>
+ *     <emsp>reverseIntakeMotor checks if the intake is not reversing, and sets the intake motor to FORWARD, then also
+ *     sets intake motor state to REVERSING</emsp> <BR>
+ */
 public class HzIntake {
 
     public DcMotor intakeMotor = null;
