+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TestingOpModes/HzAutonomousBlueOuterAsync.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.GameOpModes;
+package org.firstinspires.ftc.teamcode.TestingOpModes;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.trajectory.Trajectory;
